feat: added azure web app infrastructure
