+++ src/modules/70-pipeline/components/AbstractSteps/StepWidget.tsx
@@ -47,7 +47,6 @@ export function StepWidget<T = unknown, U = unknown>(
   formikRef: StepFormikFowardRef<T>
 ): JSX.Element | null {
   const step = factory?.getStep<T>(type)
-  console.log(step, factory)
   if (!step) {
     return __DEV__ ? <Text intent="warning">Step not found</Text> : null
   } else if (stepViewType === StepViewType.InputVariable && !step.hasStepVariables) {
