+++ src/modules/70-pipeline/components/PipelineSteps/PipelineStepInterface.ts
@@ -71,6 +71,7 @@ export enum StepType {
   ZeroNorth = 'Security',
   KubernetesAzure = 'KubernetesAzure',
   SshWinRmAzure = 'SshWinRmAzure',
+  AzureWebApp = 'AzureWebApp',
   ServerlessAwsLambdaDeploy = 'ServerlessAwsLambdaDeploy',
   ServerlessAwsLambdaRollback = 'ServerlessAwsLambdaRollback',
   ServerlessAwsInfra = 'ServerlessAwsInfra',
