+++ src/modules/75-cd/components/PipelineSteps/PipelineStepsUtil.ts
@@ -27,7 +27,7 @@ export enum InfraDeploymentType {
   KubernetesGcp = 'KubernetesGcp',
   PDC = 'Pdc',
   KubernetesAzure = 'KubernetesAzure',
-  KubernetesAzureWebApp = 'KubernetesAzureWebApp',
+  AzureWebApp = 'AzureWebApp',
   ServerlessAwsLambda = 'ServerlessAwsLambda',
   ServerlessGoogleFunctions = 'ServerlessGoogleFunctions',
   ServerlessAzureFunctions = 'ServerlessAzureFunctions',
